requirements.txt
requests~=2.32.3
rich~=13.8.1
python-dotenv~=1.0.1
openai~=1.46.0